+++ src/example4.10/src/secondfile.c
@@ -3,7 +3,7 @@
 /* example library module */
 /* only 'callable' is visible outside */
 static int32_t buf[100];
-static int32_t length;
+static size_t length;
 static void fillup();
 
 int32_t callable() {
